requirements.txt
apscheduler==3.6.3
peewee==3.14.0
PyMySQL==0.10.1
schedule==0.6.0
skale.py==6.2b0
tenacity==6.2.0
